rename HDInsightClusterDetail to HDInsightAdditionalClusterDetail, other fit and finish
