+++ HdInsightSDK/src/com/microsoft/azure/hdinsight/sdk/cluster/ClusterManager.java
